docs(radio): improve variant
