+++ app/stories/pix-radio-button.mdx
@@ -1,6 +1,7 @@
 import { Meta, Story, ArgTypes } from '@storybook/blocks';
 
 import * as ComponentStories from './pix-radio-button.stories.js';
+import * as VariantTileStories from './pix-radio-button-variant-tile.stories.js';
 
 <Meta of={ComponentStories} />
 
@@ -26,13 +27,15 @@ Pour les considérer comme un seul groupe d'inputs, **il est nécessaire qu'ils 
 
 ### Variations graphiques du composant
 
-Le PixRadioButton prend toute la largeur à sa disposition.
+Le PixRadioButton prend toute la largeur à sa disposition par défaut.
 
 <Story of={ComponentStories.FullWidth} height={100} />
 
-Si le paramètre `variant` est précisé avec la valeur `tile`, le radiobutton et son label sont visuellement regroupés dans un ensemble intégralement cliquable.
+Si le paramètre `variant` est précisé avec la valeur `tile`, le RadioButton et son label sont visuellement regroupés dans un ensemble intégralement cliquable.
 
-<Story of={ComponentStories.VariantTile} height={100} />
+<Story of={VariantTileStories.VariantTile} height={100} />
+
+Voir la [documentation de la variante tile](http://localhost:9001/iframe.html?path=/docs/form-inputs-radiobutton-variant-tile--docs).
 
 ## Disabled
 
@@ -53,3 +56,4 @@ L'attribut `@isDisabled` permet de désactiver la radio en conservant la possibi
 ## Arguments
 
 <ArgTypes of={ComponentStories} />
+./pix-radio-button-variant-tile.stories.js
